Modified 3 files
- Modified files: code/scripts/init-13b.sh, code/scripts/init-70b.sh,
code/scripts/init-7b.sh

+++ code/scripts/init-70b.sh
@@ -1,9 +1,19 @@
-# Ensure git-lfs is available
-curl -s https://packagecloud.io/install/repositories/github/git-lfs/script.deb.sh | bash
-apt-get -y install git-lfs
+echo "Initializing the 70B model."
 
-# Clone the model weights to this project
-cd /project/models && git lfs clone https://$HF_USERNAME:$HF_API_KEY@huggingface.co/meta-llama/Llama-2-70b-hf
+if [ ! -f /project/models/llama-70b.nemo ]; then
+    # Ensure git-lfs is available
+    curl -s https://packagecloud.io/install/repositories/github/git-lfs/script.deb.sh | bash
+    apt-get -y install git-lfs
 
-# Convert the model weights to NeMo format
-python /opt/NeMo/scripts/nlp_language_modeling/convert_hf_llama_to_nemo.py --in-file=/project/models/Llama-2-70b-hf/ --out-file=/project/models/llama-70b.nemo
+    # Clone the model weights to this project
+    cd /project/models && git lfs clone https://$HF_USERNAME:$HF_API_KEY@huggingface.co/meta-llama/Llama-2-70b-hf
+
+    # Convert the model weights to NeMo format
+    python /opt/NeMo/scripts/nlp_language_modeling/convert_hf_llama_to_nemo.py --in-file=/project/models/Llama-2-70b-hf/ --out-file=/project/models/llama-70b.nemo
+    echo "llama-70b.nemo is ready"
+else
+    echo "llama-70b.nemo already exists. Skipping."
+fi
+
+sleep 5
+echo "70B model is initialized."
